requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis
redis==5.0.1
hiredis==2.2.3

# ML Libraries
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
joblib==1.3.2
numpy==1.24.3
pandas==2.1.3

# Data Processing
python-dotenv==1.0.0

# External APIs
requests==2.31.0
aiohttp==3.9.1

# QR Code Processing
opencv-python==4.8.1.78
pyzbar==0.1.9
qrcode==7.4.2
pillow==10.1.0

# Security
passlib==1.7.4
python-jose==3.3.0
cryptography==41.0.7

# Monitoring
prometheus-client==0.19.0

# WebSocket
websockets==12.0

# Task Queue
celery==5.3.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.12.0
flake8==6.1.0
